# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Job Listing Portal API
 * OpenAPI spec version: 0.1.0
 */
export type HealthStatusStatus =
  (typeof HealthStatusStatus)[keyof typeof HealthStatusStatus];

export const HealthStatusStatus = {
  ok: "ok",
} as const;

export interface HealthStatus {
  status: HealthStatusStatus;
}

export interface OkResponse {
  ok: boolean;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  seeker: "seeker",
  employer: "employer",
} as const;

export type JobType = (typeof JobType)[keyof typeof JobType];

export const JobType = {
  full_time: "full_time",
  part_time: "part_time",
  contract: "contract",
  internship: "internship",
  temporary: "temporary",
} as const;

export type ApplicationStatus =
  (typeof ApplicationStatus)[keyof typeof ApplicationStatus];

export const ApplicationStatus = {
  pending: "pending",
  reviewing: "reviewing",
  interview: "interview",
  accepted: "accepted",
  rejected: "rejected",
} as const;

export interface AuthUser {
  id: string;
  email: string;
  role: UserRole;
  createdAt: string;
}

export interface MeResponse {
  user: AuthUser | null;
}

export interface RegisterPayload {
  email: string;
  /** @minLength 6 */
  password: string;
  role: UserRole;
  fullName?: string;
  companyName?: string;
}

export interface LoginPayload {
  email: string;
  password: string;
}

export interface SeekerProfile {
  userId: string;
  fullName: string;
  headline: string;
  location: string;
  phone: string;
  bio: string;
  resumeUrl: string;
  skills: string[];
  updatedAt: string;
}

export interface SeekerProfileUpdate {
  fullName?: string;
  headline?: string;
  location?: string;
  phone?: string;
  bio?: string;
  resumeUrl?: string;
  skills?: string[];
}

export interface EmployerProfile {
  userId: string;
  companyName: string;
  website: string;
  industry: string;
  location: string;
  description: string;
  logoUrl: string;
  updatedAt: string;
}

export interface EmployerProfileUpdate {
  companyName?: string;
  website?: string;
  industry?: string;
  location?: string;
  description?: string;
  logoUrl?: string;
}

export interface JobCreate {
  title: string;
  description: string;
  qualifications?: string;
  responsibilities?: string;
  location: string;
  jobType: JobType;
  /** @minimum 0 */
  salaryMin?: number;
  /** @minimum 0 */
  salaryMax?: number;
  currency?: string;
  tags?: string[];
}

export type JobUpdateStatus =
  (typeof JobUpdateStatus)[keyof typeof JobUpdateStatus];

export const JobUpdateStatus = {
  open: "open",
  closed: "closed",
} as const;

export interface JobUpdate {
  title?: string;
  description?: string;
  qualifications?: string;
  responsibilities?: string;
  location?: string;
  jobType?: JobType;
  /** @minimum 0 */
  salaryMin?: number;
  /** @minimum 0 */
  salaryMax?: number;
  currency?: string;
  tags?: string[];
  status?: JobUpdateStatus;
}

export type JobSummaryStatus =
  (typeof JobSummaryStatus)[keyof typeof JobSummaryStatus];

export const JobSummaryStatus = {
  open: "open",
  closed: "closed",
} as const;

export interface JobSummary {
  id: string;
  title: string;
  location: string;
  jobType: JobType;
  salaryMin: number;
  salaryMax: number;
  currency: string;
  tags: string[];
  status: JobSummaryStatus;
  createdAt: string;
  companyName: string;
  companyLogoUrl: string;
}

export type JobStatus = (typeof JobStatus)[keyof typeof JobStatus];

export const JobStatus = {
  open: "open",
  closed: "closed",
} as const;

export interface Job {
  id: string;
  employerId: string;
  title: string;
  description: string;
  qualifications: string;
  responsibilities: string;
  location: string;
  jobType: JobType;
  salaryMin: number;
  salaryMax: number;
  currency: string;
  tags: string[];
  status: JobStatus;
  createdAt: string;
  updatedAt: string;
  companyName: string;
  companyLogoUrl: string;
  companyWebsite: string;
  companyIndustry: string;
  companyDescription: string;
}

export type JobWithCountsStatus =
  (typeof JobWithCountsStatus)[keyof typeof JobWithCountsStatus];

export const JobWithCountsStatus = {
  open: "open",
  closed: "closed",
} as const;

export interface JobWithCounts {
  id: string;
  title: string;
  location: string;
  jobType: JobType;
  salaryMin?: number;
  salaryMax?: number;
  currency?: string;
  status: JobWithCountsStatus;
  createdAt: string;
  applicationCount: number;
}

export interface ApplyBody {
  coverLetter: string;
  resumeUrl?: string;
}

export interface Application {
  id: string;
  jobId: string;
  seekerId: string;
  coverLetter: string;
  resumeUrl: string;
  status: ApplicationStatus;
  createdAt: string;
  updatedAt: string;
}

export interface ApplicationWithJob {
  id: string;
  jobId: string;
  status: ApplicationStatus;
  coverLetter: string;
  createdAt: string;
  jobTitle: string;
  companyName: string;
  location: string;
  jobType: JobType;
}

export interface ApplicationWithSeeker {
  id: string;
  jobId: string;
  status: ApplicationStatus;
  coverLetter: string;
  resumeUrl: string;
  createdAt: string;
  seekerEmail: string;
  seekerName: string;
  seekerHeadline: string;
  seekerLocation: string;
  seekerSkills: string[];
}

export interface ApplicationUpdate {
  status: ApplicationStatus;
}

export interface DashboardStats {
  role: UserRole;
  totalJobs: number;
  totalApplications: number;
  openJobs: number;
  recentActivityCount: number;
}

export type ListJobsParams = {
  q?: string;
  location?: string;
  jobType?: JobType;
  /**
   * @minimum 0
   */
  minSalary?: number;
  featured?: boolean;
};