# lib/api-spec/openapi.yaml
openapi: 3.0.3
info:
  title: Api
  version: 0.1.0
  description: Job Listing Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Create a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterPayload"
      responses:
        "200":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Sign in to an existing account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginPayload"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: Search and list job postings
      parameters:
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: location
          schema: { type: string }
        - in: query
          name: jobType
          schema:
            $ref: "#/components/schemas/JobType"
        - in: query
          name: minSalary
          schema: { type: integer, minimum: 0 }
        - in: query
          name: featured
          schema: { type: boolean }
      responses:
        "200":
          description: Jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JobSummary"
    post:
      operationId: createJob
      tags: [jobs]
      summary: Create a new job posting (employer only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobCreate"
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
    patch:
      operationId: updateJob
      tags: [jobs]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
    delete:
      operationId: deleteJob
      tags: [jobs]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
  /jobs/{id}/apply:
    post:
      operationId: applyToJob
      tags: [applications]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplyBody"
      responses:
        "200":
          description: Applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
  /jobs/{id}/applications:
    get:
      operationId: listJobApplications
      tags: [applications]
      summary: List applications for a job (employer owner only)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApplicationWithSeeker"
...
  /applications/{id}:
    patch:
      operationId: updateApplication
      tags: [applications]
      summary: Update an application status (employer only)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
...
components:
  schemas:
...
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
    UserRole:
      type: string
      enum: [seeker, employer]
    JobType:
      type: string
      enum: [full_time, part_time, contract, internship, temporary]
    ApplicationStatus:
      type: string
      enum: [pending, reviewing, interview, accepted, rejected]
    AuthUser:
      type: object
      required: [id, email, role, createdAt]
      properties:
        id: { type: string }
        email: { type: string }
        role: { $ref: "#/components/schemas/UserRole" }
        createdAt: { type: string, format: date-time }
...
    RegisterPayload:
      type: object
      required: [email, password, role]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 6 }
        role: { $ref: "#/components/schemas/UserRole" }
        fullName: { type: string }
        companyName: { type: string }
    LoginPayload:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
...
    JobCreate:
      type: object
      required: [title, description, location, jobType]
      properties:
        title: { type: string }
        description: { type: string }
        qualifications: { type: string }
        responsibilities: { type: string }
        location: { type: string }
        jobType: { $ref: "#/components/schemas/JobType" }
        salaryMin: { type: integer, minimum: 0 }
        salaryMax: { type: integer, minimum: 0 }
        currency: { type: string }
        tags:
          type: array
          items: { type: string }
    JobUpdate:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        qualifications: { type: string }
        responsibilities: { type: string }
        location: { type: string }
        jobType: { $ref: "#/components/schemas/JobType" }
        salaryMin: { type: integer, minimum: 0 }
        salaryMax: { type: integer, minimum: 0 }
        currency: { type: string }
        tags:
          type: array
          items: { type: string }
        status:
          type: string
          enum: [open, closed]
    JobSummary:
      type: object
      required: [id, title, location, jobType, salaryMin, salaryMax, currency, tags, status, createdAt, companyName, companyLogoUrl]
      properties:
        id: { type: string }
        title: { type: string }
        location: { type: string }
        jobType: { $ref: "#/components/schemas/JobType" }
        salaryMin: { type: integer }
        salaryMax: { type: integer }
        currency: { type: string }
        tags:
          type: array
          items: { type: string }
        status: { type: string, enum: [open, closed] }
        createdAt: { type: string, format: date-time }
        companyName: { type: string }
        companyLogoUrl: { type: string }
    Job:
      type: object
      required:
        [
          id,
          employerId,
          title,
          description,
          qualifications,
          responsibilities,
          location,
          jobType,
          salaryMin,
          salaryMax,
          currency,
          tags,
          status,
          createdAt,
          updatedAt,
          companyName,
          companyLogoUrl,
          companyWebsite,
          companyIndustry,
          companyDescription,
        ]
      properties:
        id: { type: string }
        employerId: { type: string }
        title: { type: string }
        description: { type: string }
        qualifications: { type: string }
        responsibilities: { type: string }
        location: { type: string }
        jobType: { $ref: "#/components/schemas/JobType" }
        salaryMin: { type: integer }
        salaryMax: { type: integer }
        currency: { type: string }
        tags:
          type: array
          items: { type: string }
        status: { type: string, enum: [open, closed] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        companyName: { type: string }
        companyLogoUrl: { type: string }
        companyWebsite: { type: string }
        companyIndustry: { type: string }
        companyDescription: { type: string }
...
    ApplyBody:
      type: object
      required: [coverLetter]
      properties:
        coverLetter: { type: string }
        resumeUrl: { type: string }
    Application:
      type: object
      required: [id, jobId, seekerId, coverLetter, resumeUrl, status, createdAt, updatedAt]
      properties:
        id: { type: string }
        jobId: { type: string }
        seekerId: { type: string }
        coverLetter: { type: string }
        resumeUrl: { type: string }
        status: { $ref: "#/components/schemas/ApplicationStatus" }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    ApplicationWithSeeker:
      type: object
      required:
        [
          id,
          jobId,
          status,
          coverLetter,
          resumeUrl,
          createdAt,
          seekerEmail,
          seekerName,
          seekerHeadline,
          seekerLocation,
          seekerSkills,
        ]
      properties:
        id: { type: string }
        jobId: { type: string }
        status: { $ref: "#/components/schemas/ApplicationStatus" }
        coverLetter: { type: string }
        resumeUrl: { type: string }
        createdAt: { type: string, format: date-time }
        seekerEmail: { type: string }
        seekerName: { type: string }
        seekerHeadline: { type: string }
        seekerLocation: { type: string }
        seekerSkills:
          type: array
          items: { type: string }
    ApplicationUpdate:
      type: object
      required: [status]
      properties:
        status: { $ref: "#/components/schemas/ApplicationStatus" }
...