# lib/api-zod/src/generated/types/applicationStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Job Listing Portal API
 * OpenAPI spec version: 0.1.0
 */

export type ApplicationStatus =
  (typeof ApplicationStatus)[keyof typeof ApplicationStatus];

export const ApplicationStatus = {
  pending: "pending",
  reviewing: "reviewing",
  interview: "interview",
  accepted: "accepted",
  rejected: "rejected",
} as const;