# lib/api-zod/src/generated/types/applyBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Job Listing Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface ApplyBody {
  coverLetter: string;
  resumeUrl?: string;
}