# lib/api-zod/src/generated/types/authUser.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Job Listing Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from "./userRole";

export interface AuthUser {
  id: string;
  email: string;
  role: UserRole;
  createdAt: Date;
}