# lib/api-zod/src/generated/types/jobSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Job Listing Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { JobSummaryStatus } from "./jobSummaryStatus";
import type { JobType } from "./jobType";

export interface JobSummary {
  id: string;
  title: string;
  location: string;
  jobType: JobType;
  salaryMin: number;
  salaryMax: number;
  currency: string;
  tags: string[];
  status: JobSummaryStatus;
  createdAt: Date;
  companyName: string;
  companyLogoUrl: string;
}