# lib/api-zod/src/generated/types/jobUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Job Listing Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { JobType } from "./jobType";
import type { JobUpdateStatus } from "./jobUpdateStatus";

export interface JobUpdate {
  title?: string;
  description?: string;
  qualifications?: string;
  responsibilities?: string;
  location?: string;
  jobType?: JobType;
  /** @minimum 0 */
  salaryMin?: number;
  /** @minimum 0 */
  salaryMax?: number;
  currency?: string;
  tags?: string[];
  status?: JobUpdateStatus;
}