# lib/api-zod/src/generated/types/registerPayload.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Job Listing Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from "./userRole";

export interface RegisterPayload {
  email: string;
  /** @minLength 6 */
  password: string;
  role: UserRole;
  fullName?: string;
  companyName?: string;
}